feat: 할일 작성 시 자동 업데이트가 안돼서 상태 끌어올림
