+++ src/components/Write.tsx
@@ -1,12 +1,10 @@
 import { useState } from 'react';
 
-type todo = {
-  id: number;
-  text: string;
-  done: boolean;
+type WriteProps = {
+  onAdd: (text: string) => void;
 };
 
-export default function Write() {
+export default function Write({ onAdd }: WriteProps) {
   const [text, setText] = useState('');
 
   const handleSubmit = (e: React.FormEvent) => {
@@ -15,29 +13,10 @@ export default function Write() {
     const trimmed = text.trim();
     if (trimmed.length === 0) return;
 
-    const newTodo: todo = {
-      id: Date.now(),
-      text: trimmed,
-      done: false,
-    };
-
-    const stored = localStorage.getItem('todos');
-    let todos: todo[] = [];
-
-    try {
-      todos = stored ? JSON.parse(stored) : [];
-      if (!Array.isArray(todos)) {
-        todos = [];
-      }
-    } catch {
-      todos = [];
-    }
-
-    todos.push(newTodo);
-    localStorage.setItem('todos', JSON.stringify(todos));
-
+    onAdd(trimmed);
     setText('');
   };
+
   return (
     <div>
       <input type="text" placeholder="할일 입력" onChange={(e) => setText(e.target.value)} value={text} />
